Update entire app
This commit introduces a huge number of changes meant to improve the
app.

- Move all app-related code to app/.
- Add the 'make app' command to run the app from the root directory.
- Data for the app lives under app/data/.
- Style sheets for the app live under app/assets/.

I completely rewrote the app to make it more readable, to get a better
sense of where it might break, and to make it more maintainable
tomorrow. Many of the html elements in it now also have classes for
easier css styling. I also added a *ton* of css styling to make it look
a bit more proffessional.

TL;DR: I built an app.